ai-service/requirements.txt
flask
ultralytics
pillow
numpy
opencv-python-headless